Code/requirements.txt
python-dotenv
langchain-google-genai
langchain
langchain-core
langchain-pinecone
sentence-transformer
pinecone

